zql 3

+++ Objects/AsyncSqlCommand.cs
@@ -4,11 +4,11 @@
     {
         public int CommandNumber { get; set; }
 
-        public decimal Progress { get; set; }
+        public double Progress { get; set; }
 
         public string Text { get; set; }
 
-        public AsyncSqlCommand(string text, decimal progress, int commandNumber)
+        public AsyncSqlCommand(string text, double progress, int commandNumber)
         {
             this.Text = text;
             this.Progress = progress;
